Remove iframe check
To be able to use this plugin also in Kopano DeskApp which uses webviews
the check if Mattermost is run in a iframe needs to be removed. The
listener for PostMessage API messages will then always be set and the
relaying of notifications can be enabled by the from the webview parent.